fix: Use tea.Println to ensure scrollback works in deploy (#26)
This is quite annoying. Bubbletea does not print above the current
viewport (makes sense, most TUIs are not intended to have log-like
features that scroll out of the viewport). To do it, you have to use
`tea.Println`.

However, since there were dynamic components above the logs that we'd
like to `tea.Println`, (spinners and progress bar) we need to render
those as dynamic components until they have finished. We're lucky they
are static once the 'phase' has passed, so we can, when they are in
their final form, use `tea.Println` for them too - that way they don't
mess with the logs.

Output below the logs is free to be dynamic, so we can continue to use
`View()` to render those.

This has had some adverse effects. Our golden files and model test
harness isn't as useful, because it only captures output from View(). We
will need to rethink our testing as a result of this change.

Also, since we are now supporting scrollback, there is no need to
capture mouse events. Conveniently, this smashes another bug where mouse
scrollwheel works to scroll logs, but trackpads do not